board 객체, 인터페이스 생성 완료.
