+++ poj/src/main/java/com/poj/repository/board/FreeBoardRepository.java
@@ -1,6 +1,6 @@
 package com.poj.repository.board;
 
-import com.poj.entity.Board.FreeBoard;
+import com.poj.entity.board.FreeBoard;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface FreeBoardRepository extends JpaRepository<FreeBoard, Long> {
