+++ poj/src/main/java/com/poj/repository/board/QuestionBoardRepository.java
@@ -1,0 +1,7 @@
+package com.poj.repository.board;
+
+import com.poj.entity.board.QuestionBoard;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface QuestionBoardRepository extends JpaRepository<QuestionBoard, Long> {
+}
